+++ src/main/java/net/brnbrd/delightful/common/block/SalmonberryBushBlock.java
@@ -46,7 +46,7 @@ public class SalmonberryBushBlock extends BushBlock implements BonemealableBlock
 
 	@Override
 	public ItemStack getCloneItemStack(BlockState state, HitResult target, BlockGetter level, BlockPos pos, Player player) {
-		return state.getValue(AGE) <= 2 ? Util.gs(DelightfulItems.SALMONBERRY_PIPS) : DelightfulItems.SALMONBERRIES.get().getDefaultInstance();
+		return state.getValue(AGE) <= 2 ? Util.getStack(DelightfulItems.SALMONBERRY_PIPS) : DelightfulItems.SALMONBERRIES.get().getDefaultInstance();
 	}
 
 	@SuppressWarnings("deprecation")
@@ -94,7 +94,7 @@ public class SalmonberryBushBlock extends BushBlock implements BonemealableBlock
 		if (!maxAge && player.getItemInHand(hand).is(Items.BONE_MEAL)) {
 			return InteractionResult.PASS;
 		} else if (state.getValue(AGE) > 2) {
-			popResource(level, pos, Util.gs(DelightfulItems.SALMONBERRIES, (maxAge ? 2 + level.random.nextInt(2) : 1)));
+			popResource(level, pos, Util.getStack(DelightfulItems.SALMONBERRIES, (maxAge ? 2 + level.random.nextInt(2) : 1)));
 			level.playSound(null, pos, SoundEvents.SWEET_BERRY_BUSH_PICK_BERRIES, SoundSource.BLOCKS, 1F, 0.8F + level.random.nextFloat() * 0.4F);
 			BlockState blockstate = state.setValue(AGE, 1);
 			level.setBlock(pos, blockstate, 2);