+++ src/main/java/net/brnbrd/delightful/common/block/SlicedMelonBlock.java
@@ -83,7 +83,7 @@ public class SlicedMelonBlock extends Block implements ISliceable {
 
 	@Override
 	public ItemStack getSliceItem() {
-		return Util.gs(this.sliceItem, 1);
+		return Util.getStack(this.sliceItem, 1);
 	}
 
 	public ItemStack getJuiceItem() {