+++ src/main/java/net/brnbrd/delightful/common/block/SlicedMiniMelonBlock.java
@@ -84,7 +84,7 @@ public class SlicedMiniMelonBlock extends MelonBlock implements ISliceable {
 
 	@Override
 	public ItemStack getSliceItem() {
-		return Util.gs(this.sliceItem, 1);
+		return Util.getStack(this.sliceItem, 1);
 	}
 
 	@Nullable