+++ src/main/java/net/brnbrd/delightful/compat/CRCompat.java
@@ -7,10 +7,18 @@ public class CRCompat {
 	public static final String CORROSION = "corrosion";
 
 	// Resistance I is backup effect, with half duration (rounded down)
-	public static MobEffectInstance getCorrosionInstance(int duration, int amplifier) {
+	public static MobEffectInstance getCorrosionOrResistance(int duration, int amplifier) {
 		return Modid.CR.effectInstance(
 			CORROSION, duration, amplifier,
 			MobEffects.DAMAGE_RESISTANCE, duration / 2, 0
+		);
+	}
+
+	// Fire Resistance is backup effect, with half duration (rounded down)
+	public static MobEffectInstance getCorrosionOrFireResistance(int duration, int amplifier) {
+		return Modid.CR.effectInstance(
+			CORROSION, duration, amplifier,
+			MobEffects.FIRE_RESISTANCE, duration / 2, 0
 		);
 	}
 }